renamed variables for better understanding.
getting the element by 'routeID' but the ajax requests creates the routeID when the post is finished. So if I'm getting element by 'routeID' in the ajax post, the routeID hasn't been created yet and therefroe shows nothing for the data-url variable when the tweet button is clicked. need to move out of ajax post, but don't know where